Create delete endpoint and mapper the use case with the configuration

+++ src/main/kotlin/org/example/ecommerceca/item/application/usecase/DeleteItemUseCase.kt
@@ -7,6 +7,9 @@ import kotlin.uuid.Uuid
 class DeleteItemUseCase(private val repository: ItemRepository) {
     @OptIn(ExperimentalUuidApi::class)
     fun execute(itemId: Uuid) {
+        if (repository.fetch(itemId) == null) {
+            throw IllegalArgumentException("Item not found")
+        }
         repository.delete(itemId)
     }
 }
